rule-agent/requirements.txt
langchain-ibm
ibm-generative-ai
# To be removed when this issue is fixed in Langchain https://github.com/qdrant/fastembed/issues/385
onnxruntime==1.19.2
flask
flask-cors
langchain
langchain_community==0.2.6
langchain_core
langchain-text-splitters
langsmith
# To be removed when this issue is merge in Langchain https://github.com/langchain-ai/langchain/issues/22972
tenacity==8.3.0
fastembed==0.3.2
chromadb
pypdf

python-dotenv
owlready2
nltk
